Simplify context menu usage and styling; deduplicate menu items
