+++ packages/core/components/Buttons/useButtonMenu.module.css
@@ -11,8 +11,8 @@
     height: 34px;
 }
 
-.button-menu li > button, .button-menu-header {
-    padding: 0 10px;
+.button-menu li > button, .button-menu li > div > a {
+    padding: 0 12px;
 }
 
 .button-menu i, .button-menu li > div {
@@ -37,14 +37,16 @@
     border-radius: var(--small-border-radius);
 }
 
-.button-menu-divider {
-    opacity: 0.25;
+.button-menu li:empty {
+    display: none;
 }
 
 .button-menu-header {
     align-items: center;
+    border-bottom: 1px solid var(--border-color);
     display: flex;
     height: 40px;
+    padding: 0 10px;
 }
 
 .button-menu-header span {
