+++ packages/core/components/DirectoryTree/test/DirectoryTree.test.tsx
@@ -20,9 +20,11 @@ import { Provider } from "react-redux";
 import { createSandbox } from "sinon";
 
 import { TOP_LEVEL_FILE_ANNOTATIONS } from "../../../constants";
-import Annotation, { AnnotationName } from "../../../entity/Annotation";
+import Annotation from "../../../entity/Annotation";
+import AnnotationName from "../../../entity/Annotation/AnnotationName";
 import { FmsFileAnnotation } from "../../../services/FileService";
 import FileFilter from "../../../entity/FileFilter";
+import FileDownloadServiceNoop from "../../../services/FileDownloadService/FileDownloadServiceNoop";
 import HttpFileService from "../../../services/FileService/HttpFileService";
 import HttpAnnotationService from "../../../services/AnnotationService/HttpAnnotationService";
 import { initialState, interaction, reducer, reduxLogics, selection } from "../../../state";
@@ -30,7 +32,6 @@ import { initialState, interaction, reducer, reduxLogics, selection } from "../.
 import DirectoryTree from "../";
 
 import styles from "../DirectoryTreeNode.module.css";
-import FileDownloadServiceNoop from "../../../services/FileDownloadService/FileDownloadServiceNoop";
 
 describe("<DirectoryTree />", () => {
     const sandbox = createSandbox();
