+++ packages/core/entity/Annotation/index.ts
@@ -1,5 +1,6 @@
 import { get as _get, sortBy } from "lodash";
 
+import AnnotationName from "./AnnotationName";
 import annotationFormatterFactory, { AnnotationFormatter } from "../AnnotationFormatter";
 import FileDetail from "../FileDetail";
 import { AnnotationValue } from "../../services/AnnotationService";
@@ -14,20 +15,6 @@ export interface AnnotationResponse {
     type: string;
     units?: string;
 }
-
-// TypeScript (3.9) raises a bizarre error if this is an enum
-// Reference issue without clear resolution: https://github.com/microsoft/TypeScript/issues/6307
-export const AnnotationName = {
-    KIND: "Kind", // matches an annotation in filemetadata.annotation
-    FILE_ID: "file_id", // a file attribute (top-level prop on file documents in MongoDb)
-    FILE_NAME: "file_name", // a file attribute (top-level prop on file documents in MongoDb)
-    FILE_SIZE: "file_size", // a file attribute (top-level prop on file documents in MongoDb)
-    FILE_PATH: "file_path", // a file attribute (top-level prop on file documents in MongoDb)
-    PLATE_BARCODE: "Plate Barcode",
-    THUMBNAIL_PATH: "thumbnail", // (optional) file attribute (top-level prop on the file documents in MongoDb)
-    TYPE: "Type", // matches an annotation in filemetadata.annotation
-    UPLOADED: "uploaded", // matches an annotation in filemetadata.annotation
-};
 
 /**
  * Representation of an annotation available for filtering, grouping, or sorting files from FMS.
