+++ packages/core/entity/FileExplorerURL/index.ts
@@ -1,4 +1,4 @@
-import { AnnotationName } from "../Annotation";
+import AnnotationName from "../Annotation/AnnotationName";
 import FileFilter from "../FileFilter";
 import FileFolder from "../FileFolder";
 import FileSort, { SortOrder } from "../FileSort";
